Add mysql and set up router
