+++ frontend/src/main.js
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import App from './App.vue';
-import router from './router';
+import router from './plugins/router';
 import store from './store';
 import vuetify from './plugins/vuetify';
