+++ frontend/src/pages/Login.vue
@@ -1,5 +1,5 @@
 <template>
   <div class="about">
-    <h1>This is an about page</h1>
+    <h1>This will be the login page</h1>
   </div>
 </template>
